rearrange code
fmt
